rag_local/requirements.txt
chromadb==0.5.3
sentence-transformers==2.7.0
python-docx==1.1.2
regex>=2023.10.3
